+++ packages/marshal/index.js
@@ -22,7 +22,7 @@ export { deeplyFulfilled } from './src/deeplyFulfilled.js';
 export { makeTagged } from './src/makeTagged.js';
 export { Remotable, Far, ToFarFunction } from './src/make-far.js';
 
-export { QCLASS } from './src/encodeToJSON.js';
+export { QCLASS } from './src/encodeToCapData.js';
 export { makeMarshal } from './src/marshal.js';
 export { stringify, parse } from './src/marshal-stringify.js';
